Refactor MatchRecyclerViewAdapter
