+++ app/src/main/java/com/uwcs446/socialsports/ui/find/FindFragment.kt
@@ -20,7 +20,6 @@ import androidx.recyclerview.widget.RecyclerView
 import com.uwcs446.socialsports.R
 import com.uwcs446.socialsports.databinding.FragmentFindBinding
 import com.uwcs446.socialsports.domain.datetimepicker.DateTimePicker
-import com.uwcs446.socialsports.domain.match.Match
 import com.uwcs446.socialsports.domain.match.Sport
 import com.uwcs446.socialsports.ui.matchlist.MatchRecyclerViewAdapter
 import dagger.hilt.android.AndroidEntryPoint
@@ -38,8 +37,7 @@ class FindFragment : Fragment() {
 
     private val binding get() = _binding!!
 
-    private val recyclerViewData = mutableListOf<Match>()
-    private val recyclerViewAdapter = MatchRecyclerViewAdapter(recyclerViewData)
+    private val recyclerViewAdapter = MatchRecyclerViewAdapter()
 
     override fun onCreateView(
         inflater: LayoutInflater,
@@ -59,9 +57,7 @@ class FindFragment : Fragment() {
 
         // Observer which updates the recyclerview when match data changes
         findViewModel.matches.observe(viewLifecycleOwner) { matchList ->
-            recyclerViewData.clear()
-            recyclerViewData.addAll(matchList)
-            recyclerViewAdapter.notifyDataSetChanged()
+            recyclerViewAdapter.updateMatchList(matchList)
 
             binding.findProgressBar.visibility = INVISIBLE
             binding.layoutMatchList.recyclerviewMatch.visibility = VISIBLE
